+++ src/main/java/nextstep/courses/course/domain/CourseRepository.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.course.domain;
 
 public interface CourseRepository {
     int save(Course course);
